backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
sqlalchemy==2.0.23
pymysql==1.1.0
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
pillow==10.1.0
captcha==0.5.0
webdriver-manager==4.0.1